DISC-101: converted more event files to ts
