--- events/tictactoeButton.js
@@ -1,13 +1,0 @@
-const { handleGameButton } = require("../lib/tictactoe/tttHelper");
-
-module.exports = {
-    once: false,
-    name: "interactionCreate",
-    execute(interaction) {
-        if (interaction.isButton() && interaction.message.interaction.commandName == "tictactoe") {
-            handleGameButton(interaction);
-        } else {
-            return;
-        }
-    },
-};
