+++ events/tictactoeButton.ts
@@ -1,0 +1,13 @@
+import { handleGameButton } from "../lib/tictactoe/tttHelper";
+
+export const interactionCreate = {
+    once: false,
+    name: "interactionCreate",
+    execute(interaction: any) {
+        if (interaction.isButton() && interaction.message.interaction.commandName === "tictactoe") {
+            handleGameButton(interaction);
+        } else {
+            return;
+        }
+    },
+};
